+++ NEWS.md
@@ -2,6 +2,8 @@ spiro (development version)
 ===========================
 
   * Fix bug in Cortex import assigning wrong units in some cases
+  * Fix bug in Butterworth filter leading to errors in rare cases
+  * Rewritten spiro_smooth() function has now different arguments. Digital filter now work with the original raw data if possible (and not with the interpolated data).
 
 spiro 0.0.3 (2022-03-15)
 ===========================